pyproject.toml
[tool.poetry]
name = "brondby-stock-tracker"
version = "1.0.0"
description = "Brøndby IF Stock Tracker with News Sentiment Analysis"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pandas = "^2.1.3"
numpy = "^1.25.2"
requests = "^2.31.0"
yfinance = "^0.2.18"
newsapi-python = "^0.2.6"
nltk = "^3.8.1"
textblob = "^0.17.1"
sqlalchemy = "^2.0.23"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
plotly = "^5.17.0"
dash = "^2.14.2"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
feedparser = "^6.0.10"
schedule = "^1.2.0"
python-telegram-bot = "^20.7"
jinja2 = "^3.1.2"
aiofiles = "^23.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
